+++ api/fe.js
@@ -1,0 +1,5 @@
+const { generateContainerScript } = require('../forward_engineering/generateContainerScript');
+
+module.exports = {
+	generateContainerScript,
+};
